perf: switch doubles to floats
float (32-bit) uses half the memory of a double (64-bit), reducing the
size of data buffers in transforms. More working data can fit in L1/L2,
improving cache utilization

Tradeoff: float has ~7 sig figs compared to ~15 in double, but
difference is invisible to human perception.

+++ main.cc
@@ -47,7 +47,7 @@ int main(int argc, char *argv[]) {
   const char *input_path = argv[pos];
   const char *output_path = argv[pos + 1];
 
-  double quality = -1;
+  float quality = -1;
   if (pos + 2 < argc) {
     try {
       quality = std::stod(argv[pos + 2]);
@@ -65,7 +65,7 @@ int main(int argc, char *argv[]) {
   }
 
   if (quality == -1) {
-    quality = algorithm == "--dct" ? 1.0 : 0.5;
+    quality = algorithm == "--dct" ? 1.0f : 0.5f;
   }
 
   int width, height, channels;
